[FIX_GET_FEES] fix test
