+++ sovtokenfees/sovtokenfees/test/test_set_get_fees.py
@@ -26,27 +26,6 @@ from stp_core.loop.eventually import eventually
 from plenum.test.txn_author_agreement.helper import check_state_proof
 
 
-def test_get_fees_when_no_fees_set(helpers, looper):
-    def _freshness_done():
-        assert audit_ledger.size == length_after + 1
-    """
-    Update config state
-    """
-    primary = helpers.node.get_primary_node()
-    audit_ledger = primary.getLedger(AUDIT_LEDGER_ID)
-    length_after = audit_ledger.size
-    primary.master_replica._do_send_3pc_batch(ledger_id=CONFIG_LEDGER_ID)
-    looper.run(eventually(_freshness_done))
-    """
-    GET_FEES
-    """
-    response = helpers.general.do_get_fees()
-    ledger_fees = response[FEES]
-    assert ledger_fees == {}
-    check_state_proof(response)
-    helpers.node.assert_set_fees_in_memory({})
-
-
 def test_set_fees_invalid_numeric(helpers):
     """
     Test set fees with invalid numeric amount.
